Symbolic link management script.
To keep packages organized create a virtualenv and 'pip install' the
requirements.txt file. All new packages should be installed with the
'-e' option. Then edit this create_symlinks script to make the symbolic
links to the source directories within the virtualenv.

It is also helpful to keep the .gitignore up to date with linked in
sources that need not be kept in the repo.